feat: add basic project structure
The following project classes have been added :
----------------------------------------------------
ConsoleOutput - a class for outputting information about the state of the game to the console.

GameLogic - a class that will contain all the logic of the game.

Controller - Layer class between GameLogic and ConsoleOutput.

Dictionary - the class responsible for choosing a random word.

WordToGuess.txt file containing words from which a random word will be selected for guessing